Plan 9 from Bell Labs Fourth Edition 2003-03-26

+++ sys/src/9/mtx/dat.h
@@ -183,7 +183,7 @@ struct
 struct ISAConf {
 	char		*type;
 	ulong	port;
-	ulong	irq;
+	int	irq;
 	ulong	dma;
 	ulong	mem;
 	ulong	size;
